requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
django-guardian==3.2.0
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
